Update Writer.php
Hi! I've used your repo to keep a record of every message sent through a contact form in a CSV file. I've changed the fopen mode from "w+" to "a+" so it creates the file if it doesn't exist but appends a new line if it does. It works for me but maybe if would be better to create a new method: the current create() with "w+" and another append() or open() with "a+".